# docs/openapi.yaml
openapi: 3.0.3
info:
  title: pxe-server API
  description: |
    Go 一体化 PXE 装机服务接口。

    所有管理接口需先 `POST /api/login` 获取 token，并在请求头携带 `Authorization: <token>`。
    客户端直连接口（装机过程自动访问）无需登录。

    完整使用说明见 docs/API.md。
  version: "1.0.0"
servers:
  - url: http://{PXE_IP}
    variables:
      PXE_IP:
        default: 192.168.10.10
        description: pxe-server 服务 IP
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...
security:
  - BearerAuth: []

# ==================== 路径 ====================
paths:
...
  /api/config/http:
    put:
      tags: [配置]
      summary: 更新 HTTP 配置
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                listen_addr: { type: string, example: "0.0.0.0:80" }
                web_root: { type: string }
                admin_user: { type: string }
                admin_password: { type: string }
      responses:
        "200":
          description: 成功

  # ---------- DHCP 多网段 ----------
...
  /api/image/{id}:
    delete:
      tags: [资源与镜像]
      summary: 删除镜像
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: 成功
...
  /api/file/{name}:
    delete:
      tags: [文件管理]
      summary: 删除上传文件
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 成功

  # ---------- 状态与日志 ----------
...